server/requirements.txt
# TODO: Include exact versions
# Aerospike dependencies
aerospike-proximus==0.3.1

# Flask framework
flask~=2.3.2
flask-cors
werkzeug
waitress

# Web Scraper
scrapy
scrapy-playwright

# RAG dependencies
llama-cpp-python==0.2.56
sentence-transformers==2.6.1
torch
einops